Alter build process so multiple file types are in one directory
