+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
 mulberry-symbols.zip
-categories.pdf
+categories
