+++ scripts/mkzip.js
@@ -7,7 +7,7 @@ VERSION = "3.3";
 // create a file to stream archive data to.
 var output = fs.createWriteStream(`mulberry-symbols.zip`);
 var archive = archiver("zip", {
-  zlib: { level: 9 } // Sets the compression level.
+  zlib: { level: 9 }, // Sets the compression level.
 });
 
 // listen for all archive data to be written
@@ -48,7 +48,10 @@ archive.append(`Mulberry Symbols version: ${VERSION}`, { name: "VERSION.txt" });
 // append a files
 archive.file("LICENSE.txt", { name: "LICENSE.txt" });
 archive.file("symbol-info.csv", { name: "symbol-info.csv" });
-archive.file("categories.pdf", { name: "categories.pdf" });
+archive.file("categories/categories.pdf", {
+  name: "categories.pdf",
+});
+archive.file("categories/categories.html", { name: "categories.html" });
 
 // append files from a sub-directory and naming it `new-subdir` within the archive
 archive.directory("EN/", "EN-symbols");
